rename VehiclesFragment to AnalyticsFragment
