+++ NokiaHslVisualisation/app/src/main/res/navigation/nav_graph.xml
@@ -16,7 +16,7 @@
 
     <fragment
         android:id="@+id/action_vehicles"
-        android:name="com.xpyx.nokiahslvisualisation.fragments.vehicles.VehicleFragment"
+        android:name="com.xpyx.nokiahslvisualisation.fragments.analytics.AnalyticsFragment"
         android:label="@string/menu_vehicles"
         tools:layout="@layout/fragment_vehicles" />
